Move the horizontal card component to shared component
