+++ app/assets/stylesheets/mobile/horizontalCardImageComponentStyles.js
@@ -1,9 +1,9 @@
 import { StyleSheet } from 'react-native';
 import { cardBorderRadius } from '../../../constants/component_constant';
 
-const HomeHorizontalCardImageComponentStyles = StyleSheet.create({
+const HorizontalCardImageComponentStyles = StyleSheet.create({
   container:{
-    flex: 1
+    flex: 2
   },
   image: {
     borderRadius: cardBorderRadius,
@@ -15,4 +15,4 @@ const HomeHorizontalCardImageComponentStyles = StyleSheet.create({
   }
 });
 
-export default HomeHorizontalCardImageComponentStyles;
+export default HorizontalCardImageComponentStyles;
